Fix historical Elo score lookup
It turns out, it didn't work. But now it does. The problem was that the historical lookup was broken, and it was always defaulting to 1000 Elo for all drivers. Now the historical lookup is solved, so the expected outcomes should work correctly.

I also increased the points awarded for each head to head match, so that the scores can vary more than just a few hundered points from the 1000 start.